pyproject.toml
[project]
name = "netwho"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.23.0",
    "apscheduler>=3.11.1",
    "loguru>=0.7.3",
    "openai>=2.9.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "supabase>=2.25.0",
    "tenacity>=9.1.2",
]
